pyproject.toml
[project]
name = "scholarscout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.3",
    "langchain-core>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=0.2.0",
    "ollama>=0.6.0",
    "openai>=1.30.0",
    "psycopg2-binary>=2.9.0",
    "pymilvus[model]>=2.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "sentence-transformers>=2.0.0",
    "tavily-python>=0.7.24",
    "tqdm>=4.66.0",
]
